refactor: Refactor api calls
