+++ src/services/apiClient.axios.ts
@@ -2,13 +2,11 @@ import axios from "axios";
 
 // const API_BASE_URL = process.env.REACT_APP_API_BASE_URL;
 // const AUTH_TOKEN = process.env.REACT_APP_AUTH_TOKEN;
-const API_BASE_URL = "https://equinox-salt-addition.glitch.me";
-const AUTH_TOKEN = "d2lyZWFwcHMK";
-// console.log(API_BASE_URL);
-// console.log(AUTH_TOKEN);
+// const API_BASE_URL = "https://equinox-salt-addition.glitch.me";
 
-// REACT_APP_API_BASE_URL = "http://localhost:5000";
-// REACT_APP_AUTH_TOKEN = "d2lyZWFwcHMK";
+const API_BASE_URL = process.env.REACT_APP_API_BASE_URL;
+const AUTH_TOKEN = process.env.REACT_APP_AUTH_TOKEN;
+ 
 
 export const cancelTokenSource = axios.CancelToken.source();
 
@@ -21,13 +19,5 @@ export const apiClient = axios.create({
   headers: {
     "Content-Type": "application/json",
     Authorization: `Bearer ${AUTH_TOKEN}`,
-  },
-});
-
-export const productApiClient = axios.create({
-  baseURL: "https://equinox-salt-addition.glitch.me",
-  timeout: 5000,
-  headers: {
-    "Content-Type": "application/json",
   },
 });
