+++ src/services/index.ts
@@ -1,0 +1,2 @@
+export * as ProductsAPI from './product.service';
+export * as UsersAPI from './user.service';
