+++ src/ui-core/pages/HomePage/HomePage.component.tsx
@@ -1,17 +1,19 @@
-// import { Header } from "../../components";
 import { Header } from "../../components";
 import { FlashSale } from "../../templates";
 import { Category } from "../../templates";
-import { useFlashSaleProducts } from "../../../services/product.service";
+import { useQuery } from "@tanstack/react-query";
+import { ProductsAPI } from "../../../services";
+import { Item } from "types/Item";
 
 export default function HomePage(): React.JSX.Element {
+  
   //use api service to get flash sale products through react query
-  const { data } = useFlashSaleProducts();
+  const { data } = useQuery(["flashsale"], ProductsAPI.getFlashSaleProducts);
 
   return (
     <div className="home-page-container">
       <Header headingText="Modern Walk" />
-      <FlashSale flashItems={data} />
+      <FlashSale flashItems={data as Item[]} />
       <Category />
     </div>
   );
