+++ src/ui-core/pages/WomensClothing/WomensClothing.component.tsx
@@ -3,22 +3,20 @@ import { ItemCard } from "../../components";
 import { SectionLayout } from "../../layouts";
 
 import { Item } from "../../../types/Item";
-import { useWoensProducts } from "../../../services/product.service";
+import { ProductsAPI } from "../../../services";
+import { useQuery } from "@tanstack/react-query";
 
 export default function WomensClothing(): React.JSX.Element {
-  const { data } = useWoensProducts();
-  const items = data;
+  const { data } = useQuery(["womens"], ProductsAPI.getWomensProducts);
 
   return (
     <>
       <Header headingText="Modern Walk" />
       <SectionLayout heading={"Women's Clothing"}>
-        {items?.map((i: Item) => {
+        {data?.map((i: Item) => {
           return <ItemCard key={i.id} itemData={i} />;
         })}
       </SectionLayout>
     </>
   );
 }
-
-
